+++ apps/doc/tests/e2e/navigation.spec.ts
@@ -11,9 +11,7 @@ test('has title', async ({page}) => {
 	await expect(
 		page.getByRole('heading', {name: 'Fat Fuzzy Rocks'}),
 	).toBeVisible()
-	await expect(
-		page.getByRole('heading', {name: '✨ Contents ✨'}),
-	).toBeVisible()
+	await expect(page.getByRole('heading', {name: 'Contents'})).toBeVisible()
 })
 
 pages.forEach((item) => {
@@ -53,6 +51,9 @@ pages.forEach((item) => {
 
 test(`Settings menu works as expected`, async ({page}) => {
 	await page.goto('/')
+	await expect(page.getByRole('button', {name: 'Brightness'})).toBeVisible()
+	await expect(page.getByRole('button', {name: 'Contrast'})).toBeVisible()
+	await page.goto('/doc')
 	await expect(page.getByRole('button', {name: 'Brightness'})).toBeVisible()
 	await expect(page.getByRole('button', {name: 'Contrast'})).toBeVisible()
 	await expect(page.getByRole('link', {name: 'GitHub'})).toBeVisible()