+++ packages/lib/src/index.js
@@ -1,7 +1,3 @@
-import gfx from './gfx/index'
 import math from './math/index'
 
-export default {
-	gfx,
-	math,
-}
+export default {math}